[#68086384] Remove ubuntu user account
Remove the `ubuntu` user account on the first Puppet run after a new machine
has been bootstrapped. We use this to log into the the console so rarely and
there are very few problems that can't be resolved by destroying or
rebooting a machine that it doesn't justify the risk or complexity of
managing it, plus there was a bug..

New environment in Platform1 (Production and Staging) didn't have the
password hash reset from the bootstrap default. This is because the
`libshadow` gem was installed to the system Ruby but Puppet was running
under an rbenv Ruby. Without libshadow present the Puppet `user{}` type will
silently not manage passwords, so it never raised any errors.

For consistency we tried to remove the `libshadow` gem using `ensure =>
absent` but there's a bug in the `gem` provider (which we patched) that
causes it to try to uninstall from the wrong version of Ruby. So we're going
to leave that in place and file a separate bug for the ability to remove
system gems.