New self-signed certs for dev/vagrant
The old cert(s) for `*.dev.alphagov.co.uk` was issued by Comodo and expired
on 2013-11-22. It seems that nobody noticed, which means that we probably
don't need a certificate from a real CA. But it should at least be valid and
we should be prepared to make it public if it's in this repo. So replace it
with a self-signed.