[#69524064] Grafana module for graphite-1
Simple module to install our Grafana package (mirrored from our PPA),
distribute our config files into a location that doesn't muddy the directory
owned by the package, and serve them up with Nginx. This is included on
`graphite-1` since it makes sense to host it on the machine that is serving
it's dependent service (Graphite). It's HTTP because the vhost is only
exposed internally and doesn't serve anything sensitive - it will be exposed
externally as HTTPS, just like Graphite.

This means that we no longer need to clone our fork of the Grafana repo onto
the Mac Mini dashboard machine and run it using Python SimpleHTTPServer.
That's one less snowflake for 2nd line to worry about and it's usable
outside of the office.

As much as I dislike the `ensure => directory, recurse => true, source =>`
pattern to manage a whole directory of files, I think it makes sense here,
because the directory structure of `dashboards/` is flat and we'll end up
with a lot of boilerplate each time we add a new dashboard.