[#66416324] Proxy Graphite/Grafana via monitoring
Reverse proxy these vhosts externally via `monitoring-1.management` instead
of `jenkins-1.management` because it's a closer fit. I wouldn't expect to
lose access to Graphite/Grafana if I rebooted the Jenkins/Deployment
machine, but this seems a bit more acceptable if I rebooted the Nagios
machine (which depends on Graphite).

It would be nice to not make it dependent on either. We might want to still
watch graphs if we lost access to Nagios. But we can consider whether to use
a separate external IP for that as a separate story.

Include a comment above the resource that configures the reverse-proxy on
`jenkins-1` about when we can remove it. This is dependent on both this
commit, an additional SNAT rule in Preview, and DNS being updated to use
both changes.