[#69912908] Refactor custom gem prov to system_gem
Rename and refactor our custom gem provider for the package type so that:

a. It's obvious that we're overriding Puppet's default behaviour. I had
vague recollections that we'd done this, but I couldn't remember how or
where, and we shouldn't expect other people to remember this.

b. Inherit from the real provider so that we don't duplicate upstream code
and risk getting out of-step. The obvious downside to this is that our
patches *might* be broken by upstream changes. But I think that's acceptable
so that we receive other bug fixes.