NTP peer: Increase `retry_internal` and `attempts_before_hard_state`
The "check_ntp_peer" check was regularly firing as critical in our
environment but would be fixed after a period of time had passed.
After some investigation this appears to be because the remote peer is
not synchronised yet[1].

To get around this, we're going to run the check up to 3 times before
making it a hard alert and retry 5 times instead of the default 4.

[1] http://permalink.gmane.org/gmane.network.monitoring.plugins.user/5693