Don't check for truechimers
Remove the check for NTP truechimers, which is currently triggering
a few dozen critical alerts per day.

The error we see is:
```
Server not synchronized, Offset unknown, truechimers=0
```

These errors are intermittent and usually resolve after half an
hour or so.

I'm not 100% sure that it's the truechimers check that is causing this
error but I'd rather keep the other checks provided by `check_ntp_peer`
if we can.