[#70264760] Move timing of mirrorer into respective components
The timing of crawling (and artefact response times) are now part of
the govuk_mirrorer so they no longer need to be part of this script.

The respective upload timing is now part of the govuk_upload_mirror
script that gets run so it's clearer for those who are trying to
understand where the metrics are being added.