Run mirror cron job every hour
Attempt to run the mirror cron job every hour. Previously was configured
to run every 24 hours.

`setlock` prevents multiple instances of this job from starting up, so
the effect of this change is that the job should start again soon after
it finishes (no more than one hour after it last completed).

The reason for the change is that this job now takes approximately 28
hours, so by running it every hour we can at least ensure that the
mirror is as up-to-date as the script allows.

Also sets the `MAILTO` environment variable to an empty value which
will prevent cron from emailing us if `setlock` finds the lock file.

Note that this change means that the mirror job will now run almost
constantly.

Fixes [#70264126]