[#61463904] logs-cdn-1 rsyslog to receive logs
Configure rsyslog on `logs-cdn-1` to receive logs over syslog+TLS from
Fastly. We have to bind a port for each service so that we can differentiate
them and write to separate files. Most of this was cribbed from:

- https://github.com/alphagov/ci-puppet/pull/101
- https://github.com/alphagov/ci-puppet/pull/102
- https://github.com/alphagov/ci-puppet/pull/103

Certs for staging and production were added in gds/delpoyment#238. The ports
match gds/govuk-provisioning#54.