[#70371318] Add repos for ES 0.90
This is intended to be compatible with the upstream Puppet module that we
intend to switch to in the future:

- https://github.com/elasticsearch/puppet-elasticsearch

NB: Our apt module takes care of purging unmanaged apt::source resources.

I originally stuck the repo definition in the `govuk_elasticsearch` class
but this causes a dep loop. Nevermind, it probably deserves a class anyway.

    Error: Could not apply complete catalog: Found 1 dependency cycle:
    (Anchor[apt::key D88E42B4 present] => Apt::Key[Add key: D88E42B4 from Apt::Source elasticsearch-0.90] => File[elasticsearch-0.90.list] => Exec[apt_update] => Class[Apt::Update] => Package[oracle-java7-installer] => Class[Govuk_java::Oracle7::Jdk] => Exec[java-set-default-jdk] => Class[Govuk_java::Set_defaults] => Class[Govuk_elasticsearch] => Apt::Source[elasticsearch-0.90] => File[elasticsearch-0.90.list