Specify whether to mount asset-master via hiera
In [#65887346] I am removing all instances of govuk_platform and
migrating them to specific options in Hiera.

This was originally used to switch off the NFS mount of the asset master
on environments with `FACTER_govuk_platform=development` which is the
vagrant dev VM (but not vagrant-govuk, where it is set to `staging`).

By using a class parameter and setting this to `false` in
development.yaml, but `true` in common.yaml (also the class default) we
ensure that this is enabled everywhere except development.