Specify whether to use a clamav service via hiera
In [#65887346] I am removing all instances of govuk_platform and
migrating them to specific options in Hiera.

In all platforms apart from development, we install a clamav service
and use that via clamdscan to scan uploads. In development we do not
install the service and we scan via clamscan.

By using a class parameter and setting this to `false` in development.yaml,
but `true` in common.yaml (also the class default) we ensure that this is
enabled everywhere except development.