Change Transition's default db collation to utf8_unicode_ci
See https://www.pivotaltracker.com/story/show/67925580

Some time in late February the transition_production database
default collation was changed from utf8_unicode_ci to utf8_general_ci.
All tables created before then (except when we explicitly specified
otherwise in migrations) have the unicode collation, and all tables
created since then have the general collation.

We noticed this after replicating data from preview around the middle
of March. We now get a diff in the db/structure.sql file in transition
after running any rake task, which reflects these differences in collation
between the committed file and the development database as it was
replicated. This diff has been accidentally committed once and gets in
the way the rest of the time.

The reason for this change is somewhat unclear. We think it may have
been caused by the change to use the puppetlabs/mysql module
( https://github.gds/gds/puppet/pull/1443 ) which was deployed the
day before the first affected table was created, but can't see anything
in that change which might have had this effect.

We would like all the existing tables to have the unicode collation
(except for hits and host_paths which are utf8_bin) and the database
default to be utf8_unicode_ci so that new tables in the future are
created with this collation without us having to remember to specify
it in migrations. The change here will take care of the latter, and we
will have to alter the existing tables manually.

In practice we don't often order by text fields which are likely to
contain UTF8 characters within the transition app, and particularly
not for the tables which now have the general collation, so for our
users this change should have no visible impact. utf8_unicode_ci is
in principle more correct, but our main concern here really is to have
development environments which are consistent with production and to
not have the diff floating around getting in the way.