[govuk_platform] remove from mongodb::server
We deploy mongodb in two modes:

1. With `development == true`, no replicaset
2. With `development == false`, a replicaset named `production`

Previously the replicaset name was set to $govuk_platform which is
always `production` outside the dev VM (where `$development = true`)

This commit removes the ability to set the replicaset name, it's fixed
at `production` if a replicaset is actually used. Whether a replicaset
is used is set by the $development flag.

Even when we only deploy one machine in production (e.g.
exception-handler-1) we deploy it as a replicaset, so this shouldn't
cause any problems.