[govuk_platform] asset_pipeline global disable in hiera
This uses an explicit hiera lookup, which we generally try to avoid
(hence why it needs to be added to a whitelist in check_hiera.rb).
However in this case it's being used to avoid lots of duplication (I
would need to extend every class which used the `govuk::app` defined
type and have a parameter for each in hiera)

We don't ever want the asset pipeline to be enabled on a development VM.
This commit accomplishes that.