Add govuk_crawler_worker class and rabbitmq permissions class
This follows a similar pattern to other applications where a
separate class is used to define the "database" credentials.

The `$amqp_pass` is provided here as a variable so that it can
be injected using hieradata. Please see gds/deployment#270 for
the pull request where the password is set for our different
environments.