Revert "[#66672712] Mongodb should start after networking"
This reverts commit d9b33d6741cceb09fc16bd1431780ccb906e5c4e.

This doesn't seem to have improved things.  It's possible it made the
problem worse.  With this in place, mongodb seems to consistently fail
to start on reboot.  I tested reverting this change manually, and
mongodb did start on boot.

This also brings the init script more in line with the version shipped
in the 10gen package.