Create another icky top-level hiera var for gemfury
This satisfies the requirement that the mirror module should blow up
when it doesn't have a valid gemfury URL, but still passes the tests.
Unfortunately it does that by using an explicit hiera lookup - this is
bad, but it's known badness and something we can fix later _after_ we've
only got hieradata to worry about.