Remove bouncer 501 monitoring check
This check existed so that we could find out about mappings which redirected to
a host not on our whitelist. We could then add it, and deploy Bouncer with an
updated whitelist.

Transition/Bouncer have been changed so that it isn't possible to add mappings
not on the whitelist, and the whitelist can be added to by users with the
'admin' permission.

Bouncer will still return 501s for redirects not in the whitelist, but this
should not happen: it would indicate a different problem, or possibly an attack
where the attacker has inserted malicious New URLs but not managed/known to
insert whitelist entries.