Add a safety valve to prevent no users in hiera
If we don't set users::user_groups in hiera for some reason or it can't
be found, we could end up with no user accounts on the machine. That's
bad, so we should fail.

To prevent this, user_groups is a mandatory option to the users class
and in order to add no users to the machine, you must deliberately set
the following in hiera:

    users::user_groups:
        - 'none'