Switch certs based on certtype with explicit lookups
The former code relied on hiera lookups on an interpolated string which
our linter fails to grasp. Rather than fix the linter, I've collapsed
the code which switches certs with the code that checks for valid
certtype parameter. This isn't pretty, but it's no more ugly than the
former solution.