Fix spec s_whitehall_backend on non-puppetised machines
Because this node includes the mirror, which uses the concat module,
there will be a test failure if you run the specs on your mac rather
than on a linux machine which has the concat module and it's facts
installed.

This deliberately sets concat_basedir in the spec (as per s_monitoring
spec) to stop the test failure.