Remove all extdata for protect_frontend_apps
This is very very icky and still leaves quite a bit of technical debt.

I've removed the global extdata flag for protecting frontend apps which
was used in various nasty nasty places. I've replaced it with 5 separate
flags, on s_bouncer, s_calculators_frontend, s_frontend,
s_whitehall_frontend, s_licensify_frontend and one on
performance_platform::spotlight

That may not seem like an improvement, however as per my notes in the
story I have created (#73421574) I think the ones on s_frontend,
s_calculators_frontend, s_whitehall_frontend and spotlight can be removed once we are
off the old Preview environment at which point it should not be possible
to contact the frontend load balancer directly and we would be protected
by basic authentication on the cache nodes. That would then leave the
flags on s_bouncer and s_licensify_frontend which I think is reasonable.

My brain hurts and this PR needs a lot of testing by me tomorrow. I'm
only doing it this evening so I can slack off a bit tomorrow, take my
time with testing and cuddle the puppy when my brain hurts.

I'm sorry for anyone that reads this, now and in the future...