Use ensure_packages in base/packages.pp
This ensures (heh!) that if I install the same package anywhere else in
the manifest using the ensure_packages() function, they do not conflict.
Otherwise we have a race condition where if a package is installed with
ensure_packages() and then later in compilation gets defined here in a
plain packages{} resource, we will conflict.

Conflict is bad, yo.