Upgrade aptly to 0.6
We're currently using 0.4.1 in all environments because that's what the
latest was when the machines were provisioned. 0.5.1 has a fix for the
Jenkins repo which we want to mirror in [#70083936].

All releases claim to be backwards compatible. I've tested this out locally
and it seems to be fine - existing mirrors are still there.