[#70447906] Mirror RabbitMQ apt repo
Their repo is a bit weird in that they only ever present the latest version.
So we're going to have to upgrade to (presently) 3.3.4-1 once we've mirrored
it. After that we'll be able to stick with that version or pull in newer
versions as they get released.

NB: Need to run update commands per opsmanual.