Abort if hostname/domain/fqdn facts are absent
We have manifests which rely on hostname related variables from facter -
hostname, domain, and fqdn. If these aren't set because `resolv.conf(5)` is
badly configured then Puppet will happily use an empty string. Instead we
should abort the Puppet run and fix the underlying problem.

This should prevent something that Cal and Andrew experienced while bring up
Carrenza Preview when some newly provisioned machines didn't get a
domain/fqdn because the VDC network was missing some configuration for
sub-domain.

We have a resource that does `@@sshkey { $::fqdn: }` that was evaluated and
exported to PuppetDB as `Sshkey[""]`. When another machine did the same
thing the run failed complaining that there was already a resource called
`Sshkey[undef]`.

I suspect there are other more subtle misconfigurations that would also have
occurred had it got past that more obscure error.

NB: The `strict_variables` setting may also help this but it's not available
in the version of Puppet that we're currently using.