Use govuk_setenv for everything in the Procfile
All the frontend apps will need this to work behind the router on the
dev VM.  It's also good to be consistent about how we run these.

The only remaining lines that don't use `govuk_setenv` are for apps that
aren't configured in puppet and therefore don't have a corresponding
`/etc/govuk/<appname>` directory.