Remove case statements for Ubuntu Lucid
Preview on Skyscape's Interim platform is now dead. From now on we only have
Precise machines. So remove all case statements that provided backwards
compat.

It's not valuable to preserve any of these for when we move to the next LTS
version because we'll probably need to re-evaluate a whole new load of
incompatibilities.