[#70447906] Use RabbitMQ mirror and change version
Use our mirror of the RabbitMQ APT repo. We need to change the version
because, as described in eab9030, we only have one version available.

I've moved both keys to the common hiera file and prevented the use of the
upstream repo because it's never likely to work so long as we care about the
version.