[#70083936] Use Jenkins mirror and pin version
Use our mirror for the Jenkins APT repo.

We need to pin the version because, as described in 84cd312, we only have
one version available. The version chosen is what's currently on apt-1. It
will cause a downgrade from 1.554.3 which was installed automatically this
morning, but doesn't contain any important updates:

http://jenkins-ci.org/changelog-stable

This means that we'll no longer receive automatic updates like security
fixes. However we'll still be able to install the same version when upstream
release a new one and we won't fail Puppet runs everytime their mirror goes
down.