Raise urgency of a varnish not-responding alert
It is possible, when a machine is under high load, for the varnish child
process to die, being replaced by another child process that also dies,
leaving a parent process and no children. In this state, varnishd will
not respond to connections.

It will also cause 1/3 of the requests to our origin to fail.

This check has a good history of not alerting unless there is a real
problem. It alerted today on cache-3 for the first time ever and there
was a real problem.

I have:

1. Raised the priority of this alert to urgent, which will cause a page
2. Documented the resolution in the Ops Manual
3. Linked to an appropriate graph that I used to diagnose the issue.