Fix RabbitMQ on the dev vm.
It was including the rabbitmq module directly as opposed to
govuk_rabbitmq, so was missing the apt repo

This moves some of the hieradata from rabbitmq.yaml up to common.yaml so
that the dev VM gets the same config.  I've also removed some of the
values that were set to the defaults.