Replace `govuk_mirrorer` w/ `govuk_seed_crawler`
Replace the `govuk_mirrorer` gem[1], which used to index and crawl the
site, with the new `govuk_seed_crawler`[2] gem that seeds a list of URLs
for the new GOV.UK Crawler Worker.

Note that `govuk_seed_crawler` actually depends on `govuk_mirrorer` but
we no longer need to add it explicitly using Puppet as it will be pulled
in by bundler.

[1]: https://github.gds/gds/govuk_mirrorer
[2]: https://github.gds/gds/govuk_seed_crawler
[3]: https://github.com/alphagov/govuk_crawler_worker