mirror_root directory should be created by worker
The `mirror_root` directory should be created by the crawler worker
since that's the application that needs to write to that directory.

Set the directory permissions to `0755` so that the synchronisation cron
job can read from it.

I'm not keen on hardcoding the `deploy` user here but we've done it in
several other places so I think that's a job for another time.