[#71305918] Ensure `rubygems` package is absent
This commit explicitly ensures that the Rubygems package is absent. With the
version of Ruby we set in the `71305918-update-ruby-version` branch being
installed from our standard upstream repositories, Rubygems should be
automatically installed as part of the `ruby1.9.1` package.

By setting the actual Rubygems package to be absent, we can remove it from
all machines, and then entirely revoke this whole class at a later date.