Add cron job to seed crawler every 24 hrs
Configure a cron job to seed the new GOV.UK Crawler Worker every 24
hours. This is a temporary solution to ensure that the Crawler Worker
has enough to go on until our CMS applications start publishing to the
RabbitMQ exchange that the seeder publishes messages to.

Adds a lock file for this job to ensure we don't overload the exchange
with messages, and remove a redundant comment line.