Upgrade redis module to fix warnings
This upgraded Redis module fixes some warnings. By setting the version
to the one we already have installed, this is pretty much a NOOP:

Because of [thomasvandoren/puppet-redis#12](https://github.com/thomasvandoren/puppet-redis/issues/12)
the init script has changed slightly:
```
 root@development:/var/lib/puppet# diff clientbucket/8/c/4/a/d/0/2/1/8c4ad02131b58203b8ed587c8a5fb4e7/contents /etc/init.d/redis_6379
 13a14
 # chkconfig: 2345 95 20
```

Because of a [missing - to chomp empty lines in erb at L84](https://github.com/thomasvandoren/puppet-redis/commit/663b654695dba28cbd513d41120ea1e2a3039f78#diff-6f9efabb8b07e8f586ae0a23dfdbdc1bR85)
```
 root@development:/var/lib/puppet# diff clientbucket/d/b/d/3/c/4/9/4/dbd3c494397a6b84fce145e431c7d7e5/contents /etc/redis/6379.conf
 81a82
 >
```

To avoid a restart of the service, it's probably worth manually
correcting the current files in place on Production.

This paves the way later to organise an upgrade of Redis, although this
module is slightly weird in how it installs redis so we'll have to do a
lot of testing of the upgrade behaviour.