[#63135858] Enforce govuk::user name policy
Codify the discussion that usernames should match the concatenated parts of
a persons "full" name by comparing it against the "fullname" param and their
email address. This is the closest we can get to matching the LDAP policy.

This seems to be the only way we'll get this story out of our backlog. I
think it has to be enforced by code, because we typically encourage new
starters to add their own accounts to Puppet, and there's no way of them
knowing the policy or us reliably enforcing it in pull requests.

It should fail early in CI tests (rather than failing when it reaches a
Puppet master) because the `modules/users/spec/classes/users_spec.rb` test
invokes all of the available (human) user groups.

There is a whitelist of exceptions for system users and users that had
accounts before we made this decision (which won't be renamed because it's
too hard).