Add a site_root argument to `seed-crawler`
Add a `site_root` class parameter to be used as an argument passed to
the `seed-crawler` cronjob.

As a class parameter, can set this per-environment in Hiera.

Also, split the seeding and synchronisation cron jobs such that we can
enable them separately. This will allow us to seed the crawler in
Preview without synchronising Preview's content to the external mirrors.