Add a parameter to hide health check paths.
Some apps need to keep their health check information exposed, at least
to the nginx proxy level, because the vShield load balancers use it to
decide when to take a machine out of rotation. Other apps need to hide
the information from public access, because it provides an uncached,
unauthenticated point of access to some of our backend applications.

The default is to expose the health check information, to avoid
accidentally breaking compatibility. Hopefully we will be able to flip
this later to hide the endpoint by default.

The intention is to introduce this for Publisher as soon as it has a
dedicated health check endpoint. Currently, Publisher has its health
check path set to `/`; hiding this would be doubleplus ungood.