Increase RabbitMQ `vm_memory_high_watermark` to 80%
The `vm_memory_high_watermark` is a fractional value. It tells
RabbitMQ at which fraction of (total) memory usage to start
aggressively writing queued messages to disk or to use up the page
file. Whilst it does this, it also blocks publishing from client
connections. This watermark is triggered across the cluster, not just
per-node.

Our RabbitMQ nodes are dedicated to running RabbitMQ. The default
value for `vm_memory_high_watermark` is 40% (1.5GB of RAM on a 4GB RAM
VM). We should increase this value so that our cluster continues to
operate for much longer as it has the memory to do so.

For more information on this, see:

 - https://www.rabbitmq.com/memory.html#memsup
 - https://www.rabbitmq.com/configure.html#config_items