Proper nfs configuration
a new syntax for declaring NFS mounts was introduced in vagrant
1.4.0[1].

The old syntax of "nfs: true" was deprecated in 1.5.0, and no longer
recognised in newer versions. This was causing an error because the nfs
export was not getting set up on the mac side, but vagrant was
attempting to mount it all the same.

The symptom was the following output on `vagrant up`

```
...
==> default: Restarted DNS Service
==> default: Setting hostname...
==> default: Configuring and enabling network interfaces...
Guest-specific operations were attempted on a machine that is not
ready for guest communication. This should not happen and a bug
should be reported.
```

This change uses the newer syntax for vagrant versions where it's
available.

[1] https://github.com/mitchellh/vagrant/commit/850c9ea7a6be538f7ce9788f64c7437642c6de4e