Migrate puppetdb to using the postgresql module
This commit migrates puppetdb to using the Postgresql module.

NB: It is _not_ a NOOP on the puppetmaster-1 machine, because it's going
to restart postgresql. From my limited testing on the Puppetmaster in
vagrant-govuk this works, however I'm not sure what the interaction
would be with another machine running puppet at the same time, so it may
be worth manually running on the puppetmaster machines after deploy.