Refactor health check check into tested pieces.
The simplest way to make this code importable, without renaming the
script to have a `.py` extension and break consistency with all the
other check scripts, is to add a symlink.

The test script isn't automatically run from Jenkins, but hopefully
it'll prove useful to the next person who comes along to edit the check
script. Especially if they're me.