Improve govuk_check_user_naming errors
This successfully caught a name mismatch (yay!) in:

https://ci-new.alphagov.co.uk/job/puppet-branches/13314/console

However because the fullname was "Tim Mower" it reported that the username
should have been "timmower" (incorrect). But because the email address was
"timothy.mower@" it would have then failed again reporting that the username
should be "timothymower" (correct).

Shortcircuit this by comparing the fullname to the email address and *then*
comparing the username to both of those. You may still get two errors if all
three fields are different, but it won't suggest that the username should be
something that it shouldn't.