Include postgresql::lib::devel on vms running transition
This replaces the explicit package install of libpq-dev on the
development vm with an include of postgresql::lib::devel within the
definition of the transition app.

This means that all machines running the transition app will include the
necessary libraries needed to compile the pg gem needed to connect to
postgresql.