Support VMWare Fusion as a hypervisor
I switched to VMWare for my puppet VM earlier. Constantly switching
hypervisors causes problems with networking conflicts on the host, so I
would like to use VMWare Fusion for both sets of Vagrant VMs I routinely
use.

This commit adds support for bringing up these VMs with VMWare Fusion.

The update to the box version is to match the version we use for our
puppet repo and is the version for which we create Virtualbox and VMWare
Fusion images.