Set alternatives where we install packages
There is a race condition currently, where sometimes we try to set our
editor to be `vim-nox` before this package has installed. The same could
possibly happen with the `ruby1.9.1` package.

By moving the setting of the alternatives into the same class as we
install the packages and making the alternative explicitly require the
package, we remove this race condition (it's also easier to spot the
dependency if everything happens in one place).