Include pass requests in Grafana edge dashboard
When we deployed a change this morning that caused more requests to be
passed directly to origin instead of cache the Grafana dashboard appeared to
show a drop in total traffic to origin. This is because we weren't graphing
requests that go through vcl_pass. Add this to the list of metrics to fill
in the gap.

https://grafana.production.alphagov.co.uk/#/dashboard/file/edge_health.json