Coerce our postgres version to be 9.1
Unless we set this, the way the postgresql::server module looks up the
version of postgres to use will bite us when running tests on a Mac
because it's not a supported platform for the module.

By setting directly, we lose the ability to change it if we use a newer
OS version, but we can address that when necessary with a
govuk_postgresql::server parameter and some hieradata.