Remove website_host hosts(5) entries
This hiera key is currently set to the following values:

- Dev: www.dev.gov.uk
- Preview: www.preview.alphagov.co.uk
- Staging: www-origin.production.alphagov.co.uk
- Production: www.gov.uk

Removing it won't have any effect on preview and staging because we already
construct these from $app_domain. It will remove the `www.gov.uk` entry from
production which is desired because:

- we want govuk_crawler_worker to hit the CDN and take advantage of caching,
  as it does in preview, instead of hitting origin directly
- it will prevent confusion about what `www.gov.uk` actually resolves to in
  each of the environments and make production/staging identical

I am reasonably confident that nothing else depends on this.

Merging this depends on the Rate-Limit-Token changes in:

- 35581af1f3d124fd3854e5381726c1da188d520a (deployed)
- gds/deployment@bdedf3bd339f2b7dc63a0be6326686f4dcc2bed2 (deployed)
- alphagov/smokey#104 (pending at time of writing)