[#50426159] Bypass rate limiting with token header
Allow clients to bypass our rate limiting by providing the header
Rate-Limit-Token with a pre-shared key. This will allow us to:

- remove the True-Client-IP hack in Smokey which allows it to trick rate
  limiting when talking directly to origin (but not edge)
- remove the hosts(5) entry in production for `www.gov.uk` that allows
  Smokey to work that way, in order to:
  - have the mirrorer and crawler hit edge instead of origin, which will
    allow them to run faster without increasing load on origin
  - reduce confusion about where `www.gov.uk` goes in which environment
- increase the frequency of edge's health checks against origin [#73301282]

We'll need to create new "tokens" for each of the following uses and adapt
them to pass the token headers on:

- Fastly www.gov.uk service
- Smokey
- govuk_mirrorer
- govuk_crawler_worker

Although pre-shared keys aren't ideal, this seems like the least "magic"
solution to an ongoing problem. By having a separate key for each (and
commenting who it belongs to in hiera) we can at least revoke them
individually should one be leaked.

Theoretically referencing a `map{}` variable on every request will have *some*
performance impact on Nginx. But it's quite hard to measure, and my gut
feeling says that we don't serve enough traffic (even after the rate limit)
for it to matter.