Make pg_hba.conf rule optional in govuk_postgres::db
We don't need this for mapit, or for development or test databases.
Even though having the rule present in those cases shouldn't break
authentication, it would look as though it's necessary and being
relied on, especially to people who are new to postgres and the fact
that ordering of rules is really important in this file.