[#76656336] Convert kibana2_url to kibana3_url
To use the logstash.json templated dashboard as described here:

- http://www.elasticsearch.org/guide/en/kibana/current/templated-and-scripted-dashboards.html#templated-dashboards-.json

The two params that we now care about are `query` and `from`. From is a
string representation rather than number of seconds. We no longer need to
add any defaults. There are a couple of oddities, as documented by the
comments and tests:

- we use `URI.encode_www_form` to easily convert a hash to URL encoded query
  params, but also need to convert spaces from `+` to `%20` as we're not
  actually sending it as form data.
- we disallow double quotes in search terms because of a bug in Kibana.