Set up transition-postgres for testing in Preview
Transition's postgres-master branch will be deployed to Preview as
transition-postgres for a while so that we can test it thoroughly
in a production-like environment before merging it and deploying
to Production.

This renames govuk::apps::transition::postgresql_db to
govuk::apps::transition_postgres::db - the postgresql
transition_production database which has already been created in
Preview will be used by the transition-postgres app for now. We may
need to do something different when we come to switch transition to
postgresql in Production.