[#76786302] Reset version of rubygems for system ruby
This properly resets the version of Ruby to how it should be. That is:

- Gems in /var/lib/gems
- No special rubygems.rb in /usr/local/lib/site_ruby
- /usr/bin/gem1.9.1 as provided by the ruby1.9.1 package

The reason for doing this is that when removing our force-upgrade of
rubygems to 1.8.24, we failed to clean up. Nodes created since ~15th
July have rubygems 1.8.11 (as they should), but all the nodes created
prior to this have rubygems 1.8.24 still.