Ensure an http_5xx counter exists before registering a check for it
We get a lot of UNKNOWN alerts for counters which don't exist or which
don't have any data.  This is an attempt to ensure a statsd counter
exists before creating a check for it.

This is a little magical.  Because the http_5xx counters are created by
carbon-aggregator, we actually create an http_500 counter that will
cause carbon-aggregator to create a corresponding http_5xx counter.