Don't try to create counters if statsd isn't running
After #2030, @alextomlins pointed out that the dev VM doesn't have
statsd running, which meant that these execs ran every single time.

This is the nicest solution I could come up with.  The exec only runs if
both the `onlyif` and the `unless` condition are fulfilled: ie only if
statsd is running, and the counter doesn't exist.