Don't pin Ruby version for Vagrant machines
In #71305918, we pinned the version of Ruby in use in Hiera for Vagrant
machines alone. In order to bring about some consistency to how we install
Ruby, this commit removes the pinning.

The net result is that machines brought up in Vagrant should be running the
same version of Ruby as machines brought up in any other way.