Don't try another nginx upstream on timeout
Our nginx behaviour is unhelpful sometimes.  When our
appservers (particularly whitehall) come under heavy load and start
timing out, we retry every one of the whitehall-frontend servers for
every single request, and wait a full 15 seconds for each request,
resulting in a total of 75 seconds to return a state to the user.

I propose that if there's a timeout, or if we get a 502, 503 or 504
error which indicates a timeout further upstream, we *don't* try another
upstream because they're all probably quite heavily loaded at this
point, and just return the error to the user (and we may still serve
something from the mirror).

This will reduce thundering herd problems by not causing number of
requests to multiply by 5 when the appservers come under load.