Test whether a variable is set when using set -u
You can't just do `[ -z $VAR ]` when bash is running with the -u
(nounset) option, because `$VAR` is unset and will cause an error, even
though it's an error we want to handle.

Instead, we use the `[ -x ${VAR+x} ]` construct where `${VAR+x}` is a
parameter expansion which evaluates to null if var is unset and
substitutes the string "x" otherwise.