Clearer error message when trying to  in this repo
It's easy to confuse where you should run `vagrant up` and if you run it here the error message
is not as clear as it could be. I've seen lots of new/jnr devs and designers (myself included)
hit this problem and not immediately recognise the mistake.

I think the additional messaging makes this a little bit clearer.