Add feature flag for retrying timeouts, default to current behaviour
After discussion, consensus is that we should test this behaviour out in
staging for a few days to observe the difference.  A feature flag will
let us switch off retry-on-timeout in staging but not production.

Also, revert the changes to govuk_elasticsearch, and focus instead
purely on the loadbalancer module.