Remove workaround for upstart service provider in puppet 2.7
We added a workaround for an issue when running Puppet 2.7 and Ruby
1.9.3 which broke the Upstart service provider.

We don't use Puppet 2.7, we use Puppet 3.4 and that has been fixed in
this commit:
   https://github.com/puppetlabs/puppet/commit/55505db1dff0972b8fdbb611ca59c5fd81a1ce4b

This removes the workaround and enables the lint checking for
inheritance across namespaces.

I have tested that Puppet still works by bringing up a jumpbox with
vagrant-govuk and it didn't barf, so I think this is good.