Remove old redis_6379 SysVinit script
This was superseded in faa2b70 by `/etc/init.d/redis-server` which comes
from the redis package. The old process was stopped and the new one started,
but the init script and its runlevel hooks weren't removed. This means that
it will attempted to be started on boot - although it'll fail because the
binary has been removed.